controller_client/requirements.txt
websockets~=15.0
PyAutoGUI==0.9.54
python-decouple~=3.8
Pillow~=11.2.1
screeninfo~=0.8.1
pytest~=9.0.2
pytest-asyncio~=1.3.0
types-PyAutoGUI~=0.9.3.9
mypy~=1.15.0
black~=25.1.0
isort~=6.0.1
